fixes merge issues - moves some types into a separate file - addresses some linter/staticcheck issues - fixes bad counter
