+++ go.mod
@@ -15,4 +15,7 @@ require (
 	github.com/zclconf/go-cty v1.2.1
 	go.uber.org/zap v1.9.1
 	golang.org/x/net v0.0.0-20200625001655-4c5254603344 // indirect
+	golang.org/x/tools v0.0.0-20200809012840-6f4f008689da // indirect
+	golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1 // indirect
+	honnef.co/go/tools v0.0.1-2020.1.5 // indirect
 )
