+++ pkg/results/interface.go
@@ -2,6 +2,6 @@ package results
 
 // Store manages the storage and export of results information
 type Store interface {
-	AddResult(result interface{})
-	GetResults() interface{}
+	AddResult(violation *Violation)
+	GetResults() []*Violation
 }
